Extract min stake amount from bot to exchange class
